withdrawal helpers tests on ethereum forks (#212)
* cctp withdrawal

* op proof

* arb

* arg native test

* add code for arb erc20 test

* add op erc20 test

---------
