+++ packages/helpers/src/op.ts
@@ -183,7 +183,6 @@ export const buildFinalizeWithdrawal = async (
   network: number,
   withdrawalTx: string
 ) => {
-  const abiCoder = new AbiCoder()
   const provider = await getProvider(network)
 
   // Get receipt
@@ -209,11 +208,14 @@ export const buildFinalizeWithdrawal = async (
   const gasLimit = event.args.gasLimit // '491310' // Get it from `MessagePassed` event's gasLimit
   const data = event.args.data // ('0xd764ad0b0001000000000000000000000000000000000000000000000000000000005817000000000000000000000000420000000000000000000000000000000000001000000000000000000000000099c9fc46f92e8a1c0dec1b1747d010903e884be100000000000000000000000000000000000000000000000000038d7ea4c680000000000000000000000000000000000000000000000000000000000000030d4000000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000000c41635f5fd000000000000000000000000f5c28ce24acf47849988f147d5c75787c0103534000000000000000000000000f5c28ce24acf47849988f147d5c75787c010353400000000000000000000000000000000000000000000000000038d7ea4c680000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000b737570657262726964676500000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000') // Get it from `MessagePassed` event's data
 
-  // encode the transaction
-  return abiCoder.encode(
-    ['uint256', 'address', 'address', 'uint256', 'uint256', 'bytes'],
-    [nonce, sender, target, value, gasLimit, data]
-  )
+  return {
+    data,
+    gasLimit,
+    nonce,
+    sender,
+    target,
+    value,
+  }
 }
 
 const main = async () => {
